:lipstick: feat: responsiviness fixed
