+++ views/incinerator/incinerator-table/index.tsx
@@ -1,51 +1,24 @@
-import { Box, Motion, Typography } from '@interest-protocol/ui-kit';
+import { Box, Motion } from '@interest-protocol/ui-kit';
 import { FC } from 'react';
-import { useFormContext, useWatch } from 'react-hook-form';
 
-import { IncineratorNoAssetsSVG } from '@/svg';
-
-import { IncineratorForm } from '../incinerator.types';
 import IncineratorBurnButton from './burn-button';
 import IncineratorTableBody from './table-body';
 import IncineratorTableHeader from './table-header';
 
 const IncineratorTable: FC = () => {
-  const { control } = useFormContext<IncineratorForm>();
-  const fields = useWatch({ control, name: `objects` });
   return (
-    <Box mb="l" display="grid" gap="l" width="100%">
-      <Box>
+    <Box mb="l" display="grid" gap="l">
+      <Box overflowX="auto">
         <Motion
-          mt="l"
+          as="table"
           rowGap="l"
+          width="100%"
+          mt="l"
           borderCollapse="separate"
           borderSpacing="0 0.5rem"
         >
           <IncineratorTableHeader />
-          {fields?.length ? (
-            <IncineratorTableBody />
-          ) : (
-            <Box
-              gap="s"
-              width={['20rem', '20rem', '20rem', '51.188rem']}
-              display="flex"
-              height="29.188rem"
-              flexDirection="column"
-              justifyContent="center"
-              alignItems="center"
-            >
-              <Box p="m">
-                <IncineratorNoAssetsSVG
-                  maxHeight="7.375rem"
-                  maxWidth="6.625rem"
-                  width="100%"
-                />
-              </Box>
-              <Typography variant="label" size="medium">
-                You don’t have any assets
-              </Typography>
-            </Box>
-          )}
+          <IncineratorTableBody />
         </Motion>
       </Box>
       <IncineratorBurnButton />
